requirements.txt
numpy
Pillow
pytest
setuptools>=42
wheel
pybind11~=2.6.0
cmake
conan
ninja
scikit-build
